Initial implementation of CacheSessionPersistence
Accepts each of the following arguments to the constructor:

- `CacheItemPoolInterface` into which to write session data.
- `string $cookieName`, the name of the session cookie.
- `string $cookiePath = '/'`, the path(s) to which the cookie applies in the application.
- `string $cacheLimiter = 'nocache'`, a value controlling how and when sessions should expire. These follow the same rules as session.cache_limiter in ext-session.
- `int $cacheExpire = 10800`, when the session cookie should expire (if the cache limiter is one of public, private, or private_no_expire))
- `int $lastModified = null`, to indicate the last time the application was modified, which is used to generate the `Last-Modified` header when the session cookie is allowed to persist.

Internally, when initializing a session, the adapter attempts to pull a
matching identifier from the cache item pool; if none is available, or
all are expired, it will use empty data.

During persistence, it regenerates the session ID if:

- we have session data, but no identifier (new session)
- session data changed
- the session requested a regenerated identifier

We then persist the session data to the cache item pool, using the
session identifier, and a session cookie is created and injected into a
new response. If we have a cache limiter, we will attempt to create
response headers before returning the response.

If the original identifier is present in the cache, we remove it
from the cache before caching the current session data, preventing
session fixation.