Allow creating persistent session cookies
By default, cookies created to indicate a session cookie expire when the
browser closes or after a period of inactivity.

Sometimes, however, it's useful to have _persistent_ session cookies, so
that the user does not need to log in repeatedly. PHP allows this via
the `session.cookie_lifetime` INI setting.

This patch provides the ability to mark a session cookie as persistent.
When marked, the `cache_expire` value is used to set the `Max-Age`
directive of the cookie.

To configure it, use the `zend-expressive-session-cache.persistent`
configuration setting, passing a boolean `true` value to enable it. By
default, the value is `false`.