Use Expires instead of Max-Age when setting persistent cookie
The behavior of PHP's `setcookie()` and Swoole's
`Swoole\Http\Response::cookie()` are such that they can only set the
`Expires` value, and not the `Max-Age` value. While Swoole has
has a `rawcookie()` method, it still parses the cookie and sets only
values it understands, which do not include `Max-Age`. As such, the most
resilient way to handle this is to use the `Expires` directive instead,
which will lead to the same behavior.