Support `Session::getSessionLifetime()` when deciding to emit the Expires directive
This patch adapts `CacheSessionPersistence` to honor the new
`getSessionLifetime()` method when determining whether or not to emit an
`Expires` directive when creating and returning the `Set-Cookie` header.

When the engine itself does not have persistence enabled, the value of
`getSessionLifetime()` will be used. If that value is greater than zero,
an `Expires` directive using the value will be emitted.

When the engine has enabled persistence, the value will be used only if
it is greater than zero; otherwise, the engine's `$cacheExpire` value
will be used (as this defines the default behavior).

The best way to expire the session cookie is to send a `Set-Cookie`
header with an `Expires` directive set in the past, and this should
likely be done from a dedicated handler.